pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "matocr8d"
version = "0.1.0"
description = "A simple OCR library for Python"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "matocr8d Contributors", email = "contributors@matocr8d.dev"}
]
maintainers = [
    {name = "matocr8d Contributors", email = "contributors@matocr8d.dev"}
]
keywords = ["ocr", "text extraction", "image processing", "tesseract", "computer vision"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Linguistic",
]
requires-python = ">=3.7"
dependencies = [
    "Pillow>=9.0.0",
    "pytesseract>=0.3.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
    "mypy>=0.800",
]
test = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
]

[project.urls]
Homepage = "https://github.com/Akshay404error/OCR-library-python"
Documentation = "https://github.com/Akshay404error/OCR-library-python#readme"
Repository = "https://github.com/Akshay404error/OCR-library-python.git"
"Bug Tracker" = "https://github.com/Akshay404error/OCR-library-python/issues"
Changelog = "https://github.com/Akshay404error/OCR-library-python/blob/main/CHANGELOG.md"

[tool.setuptools]
packages = ["matocr8d"]

[tool.setuptools.package-data]
matocr8d = ["*.py"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=matocr8d --cov-report=term-missing --cov-report=html"

[tool.black]
line-length = 88
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
